tmpdir and dbdir: infer default from uname before command processing
now in the man pages the directory is printed (OS specific), no need to
explicitly name it in the help text. in addition, dbdir is passed to
albatross_client_local|bistro (so solo5-elftool is searched in dbdir as well).

this fixes albatross_client_local where no opam is installed. this is temporary
until solo5 tenders and utils are packaged by the operating system (at which
point, dbdir will only be used for the state file).