revise Unikernel_info and Unikernel_get
The old commands and replies are left as is, but are no longer constructed by
albatross_client. The new commands return different data:
- unikernel_info contains a SHA256 digest of the raw image (recorded by
  Vmm_unix.prepare, stored in Unikernel.t)
- unikernel_get returns explicitly a pair of bool and Cstruct.t

This is mostly backwards-compatible: new server will reply to old and new
client requests. A new client talking to an old server will lead to commands
that are not understood by the server.

Fixes #61