when inserting a policy, check policies above - but not the very same policy
It turns out, replacing a policy with some resources increased does not work
since d1a19fd5863b5bae45500c73da6b1e7b290c847f, because the code in
check_policies_above checks that the existing policy is a subpolicy of the new
one. This is wrong.

A test-case has been added which replaces the existing policy with one that
increases e.g. the number of VMs.