The table of interface can be sparse
Currently, we assume that the last ID of interfaces is equal to
the number of interfaces available. But the documentation on FreeBSD
and Linux say that the table of interface can be sparse. This patch
is a naive one which try any ID until we found what we want or we
discover all interfaces available.

The patch fix an issue on Linux (EMFILE) and properly close the
`nl_socket` in any situation (even if the function fails).