stats: use if_nametoindex (#125)
On Linux, the implementation did not work reliably since the indexes are sparse. We simplified the OCaml code and use the C function if_nametoindex directly in the C code. This works on both FreeBSD and Linux, and results in a leaner and less complex code base.

In addition, the netlink sockets are properly closed and freed.

Co-authored-by: Romain Calascibetta <romain.calascibetta@gmail.com>
Co-authored-by: Hannes Mehnert <hannes@mehnert.org>