Store the policies next to the unikernels on disk
The purpose is, while developing mollymawk - we discovered that the policies
are ephemeral - thus if you have some unikernels dumped on disk, and (re)start
albatross, they will be created (great!), but a policy_info command will return
nothing. The issue is that in mollymawk we don't want to duplicate the storage
effort, but rely on having the policy(ies) available - to generate intermediate
certificates that are nice.

To describe it a bit more technical:
 albatross is running a bunch of unikernels, deployed via TLS
 albatross is stopped
 albatross is started, unikernels are restarted
 mollymawk is started and retrieves the policies from albatross
   <- here, we read before this commit the empty set
 mollymawk tries to do something, and needs to create an intermediate
  certificate -- and by default pushes a VM=0, mem=0 policy
  <- here, all goes to hell. this is what we like to avoid

Still, mollymawk is a bit special, since it reads the policies only on startup.
This means an albatross with command-line policy modifications won't be updated
in mollymawk. I guess the path forward is to notice that and make mollymawk the
source of truth for policies (i.e. allow people to extract certificates for
command line usage, provide the CA facility).