Fix cancellation in `Send_stream
Before this patch a malicious client could ask for a block dump and then
disconnect. For a 40 GB block device this would result on my machine
albatrossd spending more than 1m30s of CPU time for no good.
With this patch the streaming is cancelled as soon as we can't send to
the client anymore.