albatross_daemon: add --name UNIKERNEL_NAME to the arguments
This is done by default, but albatross-client can select to pass --no-add-name
to avoid this behaviour. Also (for old unikernels not understanding --name or
already having --name passed) there's a restart handler to intercept error 64
and retry without adding the --name to the argv.

Also, startup can be modified in the restart command.